backend/requirements.txt
fastapi
uvicorn
sqlalchemy
langchain
langchain-openai
python-dotenv
PyPDF2
